recsys_ui/requirements.txt
streamlit==1.37.1
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
